\name{hla.snp.geno}
\docType{data}
\alias{hla.snp.geno}
\title{SNP genotypes of a study simulated from HapMap CEU}
\description{
	An object of \code{\link{hlaSNPGenoClass}} of 60 samples and 1564 SNPs.
}

\usage{hla.snp.geno}
\references{
	\url{http://hapmap.ncbi.nlm.nih.gov/downloads/genotypes/2010-08_phaseII+III/forward/}

	The International HapMap Consortium. A second generation human haplotype map of
over 3.1 million SNPs. Nature 449, 851-861. 2007.
}

\keyword{GWAS}
\keyword{genetics}
\keyword{datasets}
