\name{HLA_Type_Table}
\docType{data}
\alias{HLA_Type_Table}
\title{Four-digit HLA types of a study simulated from HapMap CEU}
\description{
	A data.frame object including HLA-A, B, C, DRB1, DQA1 and DQB1 loci of 60 samples.
}

\usage{HLA_Type_Table}
\references{
	A high-resolution HLA and SNP haplotype map for disease association studies in
the extended human MHC. de Bakker PI, McVean G, Sabeti PC, Miretti MM, Green T,
Marchini J, Ke X, Monsuur AJ, Whittaker P, Delgado M, Morrison J, Richardson A,
Walsh EC, Gao X, Galver L, Hart J, Hafler DA, Pericak-Vance M, Todd JA, Daly MJ,
Trowsdale J, Wijmenga C, Vyse TJ, Beck S, Murray SS, Carrington M, Gregory S,
Deloukas P, Rioux JD. Nat Genet. 2006 Oct;38(10):1166-72. Epub 2006 Sep 24.
}

\keyword{HLA}
\keyword{genetics}
\keyword{datasets}
