\name{hlaResource}
\alias{hlaResource}
\title{
	Model Resource Information
}
\description{
	To get online information about available HIBAG models.
}
\usage{
hlaResource()
}
\value{
	A table including information of genotyping platform, version, web link
and so on.
}
\author{Xiuwen Zheng}

\keyword{HLA}
\keyword{genetics}
