\name{print.hlaAttrBagObj}
\alias{print.hlaAttrBagObj}
\title{
	Summarize a ``hlaAttrBagObj'' object
}
\description{
	Summarize an object of \code{\link{hlaAttrBagObj}}
}
\usage{
\method{print}{hlaAttrBagObj}(x, ...)
}
\arguments{
	\item{x}{a model of \code{\link{hlaAttrBagObj}}}
	\item{...}{further arguments passed to or from other methods}
}
\value{
	None.
}
\author{Xiuwen Zheng \email{zhengx@u.washington.edu}}
\seealso{
	\code{\link{plot.hlaAttrBagClass}}, \code{\link{plot.hlaAttrBagObj}},
	\code{\link{print.hlaAttrBagClass}},
	\code{\link{summary.hlaAttrBagClass}}, \code{\link{summary.hlaAttrBagObj}}
}

\keyword{HLA}
\keyword{GWAS}
\keyword{genetics}
