\name{summary.hlaSNPHaploClass}
\alias{summary.hlaSNPHaploClass}
\title{
	Summarize the haplotype dataset
}
\description{
	Summarize the haplotype dataset.
}
\usage{
\method{summary}{hlaSNPHaploClass}(object, show=TRUE, ...)
}
\arguments{
	\item{object}{a haplotype object of \code{\link{hlaSNPHaploClass}}}
	\item{show}{if TRUE, print information}
	\item{...}{further arguments passed to or from other methods}
}
\value{
	None.
}
\author{Xiuwen Zheng \email{zhengx@u.washington.edu}}
\seealso{
	\code{\link{hlaMakeSNPGeno}}, \code{\link{hlaMakeSNPHaplo}}, \code{\link{hlaHaploSubset}},
\code{\link{hlaGenoSubset}}, \code{\link{summary.hlaSNPGenoClass}}
}

\keyword{HLA}
\keyword{GWAS}
\keyword{genetics}
