\name{HapMap_CEU_Geno}
\docType{data}
\alias{HapMap_CEU_Geno}
\title{SNP genotypes of a study simulated from HapMap CEU genotypic data}
\description{
	An object of \code{\link{hlaSNPGenoClass}} of 60 samples and 1564 SNPs.
}

\usage{HapMap_CEU_Geno}
\references{
	\url{http://hapmap.ncbi.nlm.nih.gov/downloads/genotypes/2010-08_phaseII+III/forward/}

	The International HapMap Consortium. A second generation human haplotype
map of over 3.1 million SNPs. Nature 449, 851-861. 2007.
}

\keyword{SNP}
\keyword{genetics}
\keyword{datasets}
