\name{hlaModelfromObj}
\alias{hlaModelfromObj}
\title{
	Build a model from ``hlaAttrBagObj''
}
\description{
	Build a model from an object of \code{\link{hlaAttrBagObj}}, which is stored in an
R object file.
}
\usage{
hlaModelfromObj(obj)
}
\arguments{
	\item{obj}{an object of \code{\link{hlaAttrBagObj}}}
}
\value{
	None.
}
\author{Xiuwen Zheng \email{zhengx@u.washington.edu}}
\seealso{
	\code{\link{hlaAttrBagging}}
}

\keyword{HLA}
\keyword{GWAS}
\keyword{genetics}
