\name{hlaModelFromObj}
\alias{hlaModelFromObj}
\title{
	Build a model from ``hlaAttrBagObj''
}
\description{
	Build a model from an object of \code{\link{hlaAttrBagObj}}, which is
stored in an R object file.
}
\usage{
hlaModelFromObj(obj)
}
\arguments{
	\item{obj}{an object of \code{\link{hlaAttrBagObj}}}
}
\value{
	None.
}
\author{Xiuwen Zheng}
\seealso{
	\code{\link{hlaModelToObj}}, \code{\link{hlaAttrBagging}}
}

\keyword{HLA}
\keyword{genetics}
