\name{plot.hlaAttrBagClass}
\alias{plot.hlaAttrBagClass}
\title{
	Plot a "hlaAttrBagClass" object
}
\description{
	To show a scatterplot of the numbers of individual classifiers and SNP
positions.
}
\usage{
\method{plot}{hlaAttrBagClass}(x, ...)
}
\arguments{
	\item{x}{an object of \code{\link{hlaAlleleClass}}}
	\item{...}{further arguments passed to or from other methods}
}
\author{Xiuwen Zheng}
\seealso{
	\code{\link{plot.hlaAttrBagObj}}
}

\keyword{HLA}
\keyword{genetics}
