\name{print.hlaAttrBagClass}
\alias{print.hlaAttrBagClass}
\title{
	Summarize a ``hlaAttrBagClass'' object
}
\description{
	Summarize an object of \code{\link{hlaAttrBagClass}}.
}
\usage{
\method{print}{hlaAttrBagClass}(x, ...)
}
\arguments{
	\item{x}{a model of \code{\link{hlaAttrBagClass}}}
	\item{...}{further arguments passed to or from other methods}
}
\value{
	None.
}
\author{Xiuwen Zheng}
\seealso{
	\code{\link{plot.hlaAttrBagClass}}, \code{\link{summary.hlaAttrBagClass}}
}

\keyword{HLA}
\keyword{genetics}
