\name{summary.hlaSNPHaploClass}
\alias{summary.hlaSNPHaploClass}
\title{
	Summarize the haplotypic dataset
}
\description{
	Summarize the haplotypic dataset.
}
\usage{
\method{summary}{hlaSNPHaploClass}(object, show=TRUE, ...)
}
\arguments{
	\item{object}{a haplotype object of \code{\link{hlaSNPHaploClass}}}
	\item{show}{if TRUE, print information}
	\item{...}{further arguments passed to or from other methods}
}
\value{
	None.
}
\author{Xiuwen Zheng}
\seealso{
	\code{\link{hlaMakeSNPGeno}}, \code{\link{hlaMakeSNPHaplo}},
	\code{\link{hlaGenoSubset}}, \code{\link{hlaHaploSubset}},
	\code{\link{summary.hlaSNPGenoClass}}
}

\keyword{SNP}
\keyword{genetics}
